+++ src/libutil-tests/meson.build
@@ -60,6 +60,7 @@ sources = files(
   'executable-path.cc',
   'file-content-address.cc',
   'file-descriptor.cc',
+  'file-system-at.cc',
   'file-system.cc',
   'git.cc',
   'hash.cc',
