+++ src/libutil-tests/unix/meson.build
@@ -1,3 +1,3 @@
 sources += files(
-  'file-descriptor.cc',
+  'file-system-at.cc',
 )
