+++ src/libutil/fs-sink.cc
@@ -2,6 +2,7 @@
 
 #include "nix/util/error.hh"
 #include "nix/util/config-global.hh"
+#include "nix/util/file-system-at.hh"
 #include "nix/util/fs-sink.hh"
 
 #ifdef _WIN32
