+++ src/libutil/windows/meson.build
@@ -3,6 +3,7 @@ sources += files(
   'environment-variables.cc',
   'file-descriptor.cc',
   'file-path.cc',
+  'file-system-at.cc',
   'file-system.cc',
   'known-folders.cc',
   'muxable-pipe.cc',
